make plugin directory and an init.lua file inside it and move the plugin configs there
