+++ .config/nvim/lua/plugins/init.lua
@@ -1,0 +1,545 @@
+-- You can add your own plugins here or in other files in this directory!
+--  I promise not to create any merge conflicts in this directory :)
+--
+-- See the kickstart.nvim README for more information
+return {
+
+  -- [[ Configure and install plugins ]]
+  -- NOTE: Here is where you install your plugins.
+  require('lazy').setup({
+
+    'NMAC427/guess-indent.nvim', -- Detect tabstop and shiftwidth automatically
+    {
+      'ayu-theme/ayu-vim',
+      lazy = false,
+      priority = 1000,
+      config = function()
+        vim.g.ayucolor = 'dark'
+        vim.cmd.colorscheme 'ayu'
+        -- Common file types:
+        vim.cmd 'highlight DevIconsLua guifg=#FFC107' -- Example: Ayu's yellow/orange
+        vim.cmd 'highlight DevIconsJavaScript guifg=#FFC107' -- Example: Ayu's yellow/orange
+        vim.cmd 'highlight DevIconsTypeScript guifg=#2196F3' -- Example: Ayu's blue
+        vim.cmd 'highlight DevIconsPython guifg=#00FFCC' -- Example: Ayu's green/cyan
+        vim.cmd 'highlight DevIconsHtml guifg=#E57373' -- Example: Ayu's red/orange
+        vim.cmd 'highlight DevIconsCss guifg=#64B5F6' -- Example: Ayu's light blue
+        vim.cmd 'highlight DevIconsJson guifg=#FF5722' -- Example: Ayu's darker orange/red
+        vim.cmd 'highlight DevIconsMarkdown guifg=#D0D0D0' -- Example: Neutral gray/white
+
+        -- Folder Icon:
+        vim.cmd 'highlight DevIconsFolder guifg=#8be9fd' -- Example: Ayu's light blue for folders
+        vim.cmd 'highlight link Hidden DevIconsDefault' -- This will apply the default devicon color
+        -- vim.cmd 'highlight link Hidden Normal'
+        -- Default (generic) file icon color (if .txt is already green, it might be this one):
+        -- If .txt is green, maybe DevIconsDefault is linked to a green group, or
+        -- you specifically have an override.
+        -- If you want all "other" files to be green too, you could use this:
+        -- vim.cmd("highlight DevIconsDefault guifg=#00FFCC")
+
+        -- Git Status Icons (if you want them colored by devicons):
+        vim.cmd 'highlight DevIconsGitIgnored guifg=#9e9e9e' -- Example: muted gray
+        vim.cmd 'highlight DevIconsGitUntracked guifg=#FFA000' -- Example: orange
+        vim.cmd 'highlight DevIconsGitModified guifg=#FFD700' -- Example: yellow
+        vim.cmd 'highlight DevIconsGitStaged guifg=#00FF00' -- Example: green
+        -- ... and so on for other git statuses.
+      end,
+    },
+
+    { -- Adds git related signs to the gutter, as well as utilities for managing changes
+      'lewis6991/gitsigns.nvim',
+      opts = {
+        signs = {
+          add = { text = '+' },
+          change = { text = '~' },
+          delete = { text = '_' },
+          topdelete = { text = '‾' },
+          changedelete = { text = '~' },
+        },
+      },
+    },
+
+    { -- Useful plugin to show you pending keybinds (I love this one thank you TJ)
+      'folke/which-key.nvim',
+      event = 'VimEnter', -- Sets the loading event to 'VimEnter'
+      opts = {
+        delay = 0,
+        icons = {
+          mappings = vim.g.have_nerd_font,
+          keys = vim.g.have_nerd_font and {} or {
+            Up = '<Up> ',
+            Down = '<Down> ',
+            Left = '<Left> ',
+            Right = '<Right> ',
+            C = '<C-…> ',
+            M = '<M-…> ',
+            D = '<D-…> ',
+            S = '<S-…> ',
+            CR = '<CR> ',
+            Esc = '<Esc> ',
+            ScrollWheelDown = '<ScrollWheelDown> ',
+            ScrollWheelUp = '<ScrollWheelUp> ',
+            NL = '<NL> ',
+            BS = '<BS> ',
+            Space = '<Space> ',
+            Tab = '<Tab> ',
+            F1 = '<F1>',
+            F2 = '<F2>',
+            F3 = '<F3>',
+            F4 = '<F4>',
+            F5 = '<F5>',
+            F6 = '<F6>',
+            F7 = '<F7>',
+            F8 = '<F8>',
+            F9 = '<F9>',
+            F10 = '<F10>',
+            F11 = '<F11>',
+            F12 = '<F12>',
+          },
+        },
+
+        -- Document existing key chains
+        spec = {
+          { '<leader>s', group = '[S]earch' },
+          { '<leader>t', group = '[T]oggle' },
+          { '<leader>h', group = 'Git [H]unk', mode = { 'n', 'v' } },
+        },
+      },
+    },
+
+    {
+      'nvim-tree/nvim-web-devicons',
+      enabled = vim.g.have_nerd_font,
+      lazy = false,
+      config = function()
+        require('nvim-web-devicons').setup {
+          color_icons = true,
+          default = true,
+        }
+      end,
+    },
+
+    { -- Fuzzy Finder (files, lsp, etc)
+      'nvim-telescope/telescope.nvim',
+      event = 'VimEnter',
+      dependencies = {
+        'nvim-lua/plenary.nvim',
+        {
+          'nvim-telescope/telescope-fzf-native.nvim',
+          build = 'make',
+          cond = function()
+            return vim.fn.executable 'make' == 1
+          end,
+        },
+
+        { 'nvim-telescope/telescope-ui-select.nvim' },
+
+        { 'nvim-tree/nvim-web-devicons' },
+      },
+
+      config = function()
+        -- [[ Configure Telescope ]]
+        require('telescope').setup {
+          -- defaults = {
+          --   mappings = {
+          --     i = { ['<c-enter>'] = 'to_fuzzy_refine' },
+          --   },
+          -- },
+          -- pickers = {}
+          extensions = {
+            ['ui-select'] = {
+              require('telescope.themes').get_dropdown(),
+            },
+          },
+        }
+
+        pcall(require('telescope').load_extension, 'fzf')
+        pcall(require('telescope').load_extension, 'ui-select')
+
+        local builtin = require 'telescope.builtin'
+        vim.keymap.set('n', '<leader>sh', builtin.help_tags, { desc = '[S]earch [H]elp' })
+        vim.keymap.set('n', '<leader>sk', builtin.keymaps, { desc = '[S]earch [K]eymaps' })
+        vim.keymap.set('n', '<leader>sf', function()
+          builtin.find_files {
+            cwd = '~',
+            hidden = true,
+            no_ignore = false,
+          }
+        end, { desc = '[S]earch [F]iles' })
+        vim.keymap.set('n', '<leader>ss', builtin.builtin, { desc = '[S]earch [S]elect Telescope' })
+        vim.keymap.set('n', '<leader>sw', builtin.grep_string, { desc = '[S]earch current [W]ord' })
+        vim.keymap.set('n', '<leader>sg', builtin.live_grep, { desc = '[S]earch by [G]rep' })
+        vim.keymap.set('n', '<leader>sd', builtin.diagnostics, { desc = '[S]earch [D]iagnostics' })
+        vim.keymap.set('n', '<leader>sr', builtin.resume, { desc = '[S]earch [R]esume' })
+        vim.keymap.set('n', '<leader>s.', builtin.oldfiles, { desc = '[S]earch Recent Files ("." for repeat)' })
+        vim.keymap.set('n', '<leader><leader>', builtin.buffers, { desc = '[ ] Find existing buffers' })
+        vim.keymap.set('n', '<leader>/', function()
+          builtin.current_buffer_fuzzy_find(require('telescope.themes').get_dropdown {
+            winblend = 10,
+            previewer = false,
+          })
+        end, { desc = '[/] Fuzzily search in current buffer' })
+        vim.keymap.set('n', '<leader>s/', function()
+          builtin.live_grep {
+            grep_open_files = true,
+            prompt_title = 'Live Grep in Open Files',
+          }
+        end, { desc = '[S]earch [/] in Open Files' })
+        vim.keymap.set('n', '<leader>sn', function()
+          builtin.find_files { cwd = vim.fn.stdpath 'config' }
+        end, { desc = '[S]earch [N]eovim files' })
+      end,
+    },
+    {
+      'folke/lazydev.nvim',
+      ft = 'lua',
+      opts = {
+        library = {
+          { path = '${3rd}/luv/library', words = { 'vim%.uv' } },
+        },
+      },
+    },
+    {
+      'neovim/nvim-lspconfig',
+      dependencies = {
+        -- NOTE: `opts = {}` is the same as calling `require('mason').setup({})`
+        { 'mason-org/mason.nvim', opts = {} },
+        'mason-org/mason-lspconfig.nvim',
+        'WhoIsSethDaniel/mason-tool-installer.nvim',
+        { 'j-hui/fidget.nvim', opts = {} },
+        'saghen/blink.cmp',
+      },
+      config = function()
+        vim.api.nvim_create_autocmd('LspAttach', {
+          group = vim.api.nvim_create_augroup('kickstart-lsp-attach', { clear = true }),
+          callback = function(event)
+            local map = function(keys, func, desc, mode)
+              mode = mode or 'n'
+              vim.keymap.set(mode, keys, func, { buffer = event.buf, desc = 'LSP: ' .. desc })
+            end
+
+            -- Rename the variable under your cursor.
+            map('grn', vim.lsp.buf.rename, '[R]e[n]ame')
+
+            -- Execute a code action, usually your cursor needs to be on top of an error
+            map('gra', vim.lsp.buf.code_action, '[G]oto Code [A]ction', { 'n', 'x' })
+
+            -- Find references for the word under your cursor.
+            map('grr', require('telescope.builtin').lsp_references, '[G]oto [R]eferences')
+
+            -- Jump to the implementation of the word under your cursor.
+            map('gri', require('telescope.builtin').lsp_implementations, '[G]oto [I]mplementation')
+
+            -- Jump to the definition of the word under your cursor.
+            --  To jump back, press <C-t>.
+            map('grd', require('telescope.builtin').lsp_definitions, '[G]oto [D]efinition')
+
+            -- WARN: This is not Goto Definition, this is Goto Declaration.
+            --  For example, in C this would take you to the header.
+            map('grD', vim.lsp.buf.declaration, '[G]oto [D]eclaration')
+
+            -- Fuzzy find all the symbols in your current document.
+            --  Symbols are things like variables, functions, types, etc.
+            map('gO', require('telescope.builtin').lsp_document_symbols, 'Open Document Symbols')
+
+            -- Fuzzy find all the symbols in your current workspace.
+            --  Similar to document symbols, except searches over your entire project.
+            map('gW', require('telescope.builtin').lsp_dynamic_workspace_symbols, 'Open Workspace Symbols')
+
+            -- Jump to the type of the word under your cursor.
+            --  Useful when you're not sure what type a variable is and you want to see
+            --  the definition of its *type*, not where it was *defined*.
+            map('grt', require('telescope.builtin').lsp_type_definitions, '[G]oto [T]ype Definition')
+
+            -- This function resolves a difference between neovim nightly (version 0.11) and stable (version 0.10)
+            ---@param client vim.lsp.Client
+            ---@param method vim.lsp.protocol.Method
+            ---@param bufnr? integer some lsp support methods only in specific files
+            ---@return boolean
+            local function client_supports_method(client, method, bufnr)
+              if vim.fn.has 'nvim-0.11' == 1 then
+                return client:supports_method(method, bufnr)
+              else
+                return client.supports_method(method, { bufnr = bufnr })
+              end
+            end
+
+            local client = vim.lsp.get_client_by_id(event.data.client_id)
+            if client and client_supports_method(client, vim.lsp.protocol.Methods.textDocument_documentHighlight, event.buf) then
+              local highlight_augroup = vim.api.nvim_create_augroup('kickstart-lsp-highlight', { clear = false })
+              vim.api.nvim_create_autocmd({ 'CursorHold', 'CursorHoldI' }, {
+                buffer = event.buf,
+                group = highlight_augroup,
+                callback = vim.lsp.buf.document_highlight,
+              })
+
+              vim.api.nvim_create_autocmd({ 'CursorMoved', 'CursorMovedI' }, {
+                buffer = event.buf,
+                group = highlight_augroup,
+                callback = vim.lsp.buf.clear_references,
+              })
+
+              vim.api.nvim_create_autocmd('LspDetach', {
+                group = vim.api.nvim_create_augroup('kickstart-lsp-detach', { clear = true }),
+                callback = function(event2)
+                  vim.lsp.buf.clear_references()
+                  vim.api.nvim_clear_autocmds { group = 'kickstart-lsp-highlight', buffer = event2.buf }
+                end,
+              })
+            end
+
+            -- The following code creates a keymap to toggle inlay hints in your
+            -- code, if the language server you are using supports them
+            if client and client_supports_method(client, vim.lsp.protocol.Methods.textDocument_inlayHint, event.buf) then
+              map('<leader>th', function()
+                vim.lsp.inlay_hint.enable(not vim.lsp.inlay_hint.is_enabled { bufnr = event.buf })
+              end, '[T]oggle Inlay [H]ints')
+            end
+          end,
+        })
+
+        -- Diagnostic Config
+        vim.diagnostic.config {
+          severity_sort = true,
+          float = { border = 'rounded', source = 'if_many' },
+          underline = { severity = vim.diagnostic.severity.ERROR },
+          signs = vim.g.have_nerd_font and {
+            text = {
+              [vim.diagnostic.severity.ERROR] = '󰅚 ',
+              [vim.diagnostic.severity.WARN] = '󰀪 ',
+              [vim.diagnostic.severity.INFO] = '󰋽 ',
+              [vim.diagnostic.severity.HINT] = '󰌶 ',
+            },
+          } or {},
+          virtual_text = {
+            source = 'if_many',
+            spacing = 2,
+            format = function(diagnostic)
+              local diagnostic_message = {
+                [vim.diagnostic.severity.ERROR] = diagnostic.message,
+                [vim.diagnostic.severity.WARN] = diagnostic.message,
+                [vim.diagnostic.severity.INFO] = diagnostic.message,
+                [vim.diagnostic.severity.HINT] = diagnostic.message,
+              }
+              return diagnostic_message[diagnostic.severity]
+            end,
+          },
+        }
+
+        local capabilities = require('blink.cmp').get_lsp_capabilities()
+
+        local servers = {
+          clangd = {},
+          -- gopls = {},
+          pyright = {},
+          -- rust_analyzer = {},
+          -- ... etc. See `:help lspconfig-all` for a list of all the pre-configured LSPs
+          ts_ls = {},
+          lua_ls = {
+            settings = {
+              Lua = {
+                completion = {
+                  callSnippet = 'Replace',
+                },
+              },
+            },
+          },
+        }
+
+        local ensure_installed = vim.tbl_keys(servers or {})
+        vim.list_extend(ensure_installed, {
+          'stylua', -- Used to format Lua code
+        })
+        require('mason-tool-installer').setup { ensure_installed = ensure_installed }
+
+        require('mason-lspconfig').setup {
+          ensure_installed = {}, -- explicitly set to an empty table (Kickstart populates installs via mason-tool-installer)
+          automatic_installation = false,
+          handlers = {
+            function(server_name)
+              local server = servers[server_name] or {}
+              server.capabilities = vim.tbl_deep_extend('force', {}, capabilities, server.capabilities or {})
+              require('lspconfig')[server_name].setup(server)
+            end,
+          },
+        }
+      end,
+    },
+
+    { -- Autoformat
+      'stevearc/conform.nvim',
+      event = { 'BufWritePre' },
+      cmd = { 'ConformInfo' },
+      keys = {
+        {
+          '<leader>f',
+          function()
+            require('conform').format { async = true, lsp_format = 'fallback' }
+          end,
+          mode = '',
+          desc = '[F]ormat buffer',
+        },
+      },
+      opts = {
+        notify_on_error = false,
+        format_on_save = function(bufnr)
+          local disable_filetypes = { c = true, cpp = true }
+          if disable_filetypes[vim.bo[bufnr].filetype] then
+            return nil
+          else
+            return {
+              timeout_ms = 500,
+              lsp_format = 'fallback',
+            }
+          end
+        end,
+        formatters_by_ft = {
+          lua = { 'stylua' },
+          -- Conform can also run multiple formatters sequentially
+          python = { 'isort', 'black' },
+          -- You can use 'stop_after_first' to run the first available formatter from the list
+          javascript = { 'prettierd', 'prettier', stop_after_first = true },
+        },
+      },
+    },
+
+    { -- Autocompletion
+      'saghen/blink.cmp',
+      event = 'VimEnter',
+      version = '1.*',
+      dependencies = {
+        -- Snippet Engine
+        {
+          'L3MON4D3/LuaSnip',
+          version = '2.*',
+          build = (function()
+            -- Remove the below condition to re-enable on windows.
+            if vim.fn.has 'win32' == 1 or vim.fn.executable 'make' == 0 then
+              return
+            end
+            return 'make install_jsregexp'
+          end)(),
+          dependencies = {
+
+            {
+              'rafamadriz/friendly-snippets',
+              config = function()
+                require('luasnip.loaders.from_vscode').lazy_load()
+              end,
+            },
+          },
+          opts = {},
+        },
+        'folke/lazydev.nvim',
+      },
+      --- @module 'blink.cmp'
+      --- @type blink.cmp.Config
+      opts = {
+        keymap = {
+          preset = 'default',
+        },
+
+        appearance = {
+          nerd_font_variant = 'mono',
+        },
+
+        completion = {
+          documentation = { auto_show = false, auto_show_delay_ms = 500 },
+        },
+
+        sources = {
+          default = { 'lsp', 'path', 'snippets', 'lazydev' },
+          providers = {
+            lazydev = { module = 'lazydev.integrations.blink', score_offset = 100 },
+          },
+        },
+
+        snippets = { preset = 'luasnip' },
+
+        fuzzy = { implementation = 'lua' },
+
+        -- Shows a signature help window while you type arguments for a function
+        signature = { enabled = true },
+      },
+    },
+
+    -- Highlight todo, notes, etc in comments
+    { 'folke/todo-comments.nvim', event = 'VimEnter', dependencies = { 'nvim-lua/plenary.nvim' }, opts = { signs = false } },
+
+    { -- Collection of various small independent plugins/modules
+      'echasnovski/mini.nvim',
+      config = function()
+        require('mini.ai').setup { n_lines = 500 }
+        require('mini.surround').setup()
+        local statusline = require 'mini.statusline'
+        statusline.setup { use_icons = vim.g.have_nerd_font }
+        ---@diagnostic disable-next-line: duplicate-set-field
+        statusline.section_location = function()
+          return '%2l:%-2v'
+        end
+      end,
+    },
+    {
+      'christoomey/vim-tmux-navigator',
+      lazy = false,
+    },
+    { -- Highlight, edit, and navigate code
+      'nvim-treesitter/nvim-treesitter',
+      build = ':TSUpdate',
+      main = 'nvim-treesitter.configs', -- Sets main module to use for opts
+      opts = {
+        ensure_installed = { 'bash', 'c', 'diff', 'html', 'lua', 'luadoc', 'markdown', 'markdown_inline', 'query', 'vim', 'vimdoc' },
+        auto_install = true,
+        highlight = {
+          enable = true,
+          additional_vim_regex_highlighting = { 'ruby' },
+        },
+        indent = { enable = true, disable = { 'ruby' } },
+      },
+
+      {
+        'stevearc/oil.nvim',
+        ---@module 'oil'
+        ---@type oil.SetupOpts
+        opts = {
+          show_hidden = true,
+        },
+        -- Optional dependencies
+        -- dependencies = { { 'echasnovski/mini.icons', opts = {} } },
+        dependencies = { 'nvim-tree/nvim-web-devicons' }, -- use if you prefer nvim-web-devicons
+        -- Lazy loading is not recommended because it is very tricky to make it work correctly in all situations.
+        lazy = false,
+      },
+    },
+
+    -- NOTE: Next step on your Neovim journey: Add/Configure additional plugins for Kickstart
+    -- require 'kickstart.plugins.debug',
+    -- require 'kickstart.plugins.indent_line',
+    -- require 'kickstart.plugins.lint',
+    require 'kickstart.plugins.autopairs',
+    require 'kickstart.plugins.neo-tree',
+    -- require 'kickstart.plugins.gitsigns', -- adds gitsigns recommend keymaps
+
+    -- NOTE: The import below can automatically add your own plugins, configuration, etc from `lua/custom/plugins/*.lua`
+    --    This is the easiest way to modularize your config.
+
+    { import = 'custom.plugins' },
+  }, {
+    ui = {
+      icons = vim.g.have_nerd_font and {} or {
+        cmd = '⌘',
+        config = '🛠',
+        event = '📅',
+        ft = '📂',
+        init = '⚙',
+        keys = '🗝',
+        plugin = '🔌',
+        runtime = '💻',
+        require = '🌙',
+        source = '📄',
+        start = '🚀',
+        task = '📌',
+        lazy = '💤 ',
+      },
+    },
+  }),
+}
